refactor: move users db logic to seperate module
